# docs/contracts/portal-api-v1.4.1/docs/web-portal-api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: SSPOS web portal API
  version: "1.4.0"
  summary: "The contract between SSPOS 3 tills and the web portal: sync, licensing, devices, onboarding, migration."
  description: "Contract major version 1 (X-SSPOS-Contract: 1), document revision 1.4 (package v1.4, 2026-09-29). The prose contract is docs/web-portal-api.md; where this file and a sample disagree, the sample wins. Schemas are JSON Schema 2020-12 referenced by relative path (schemas/ is generated from the till's C# types; licensing/schemas/ is the section 17 contract). Every reply may gain new fields; clients ignore fields they do not know. The till calls nine operations: sync hello / push / pull, licence activate / validate / redeem, devices/deactivate, cloud/migrate and its complete; devices/activate is kept for reference only (deprecated - not called). Every POST except a normal (delta) push carries an Idempotency-Key, one per logical request: a retry repeats the key and the first attempt's body byte for byte (section 17.11 rule 5)."
  license:
    name: Proprietary - SSPOS
    identifier: LicenseRef-SSPOS-Proprietary
servers:
  - url: "https://{portalHost}"
    description: The portal. HTTPS only (TLS 1.2+).
    variables:
      portalHost:
        default: portal.example.co.uk
...
security:
  - branchKey: []
paths:
...
  /api/v1/devices/deactivate:
    post:
      tags:
        - Devices
      operationId: deactivateDevice
      summary: Release this till - free its seat, or release the main till for a transfer (section 17.7)
      description: "Sent by the till being released (Settings > Licence > Deactivate this till), for its own registerId and installId. Branch key when the till holds one (a per-till licence is identified by registerId + installId). Idempotent by registerId (twice = the same reply) and by Idempotency-Key (a retry repeats key and body until answered). On 200 the till forgets its cloud link, withdraws the licence it holds and writes an AuditLog row; any refusal or no answer changes nothing on the till."
      security:
        - branchKey: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./licensing/schemas/deactivate-request.schema.json"
            examples:
              secondTill:
                summary: Remove Till 2
                value:
                  registerId: "01K5T0Q8C4000000000000R002"
                  installId: "01K5T0Q8C4000000000000H002"
                  reason: removed
                  note: Till 2 moved to the stock room as a spare.
              mainTill:
                summary: Move the main till to a new PC
                value:
                  registerId: "01K5T0Q8C4000000000000R001"
                  installId: "01K5T0Q8C4000000000000H001"
                  reason: transfer
                  note: New PC arriving Monday.
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./licensing/schemas/deactivate-reply.schema.json"
              examples:
                secondTill:
                  summary: Seat freed
                  value:
                    registerId: "01K5T0Q8C4000000000000R002"
                    seat: deactivated
                    seatsInUse: 1
                    maxRegisters: 2
                    apiKeyRevoked: false
                    transferCode: null
                    transferCodeExpiresAt: null
                    portalTimeUtc: "2026-10-04T17:00:01Z"
                    messages: []
                mainTill:
                  summary: Released, transfer code issued
                  value:
                    registerId: "01K5T0Q8C4000000000000R001"
                    seat: deactivated
                    seatsInUse: 1
                    maxRegisters: 3
                    apiKeyRevoked: true
                    transferCode: SSP-TRF-8M3K-2QXD
                    transferCodeExpiresAt: "2026-10-11T17:00:00Z"
                    portalTimeUtc: "2026-10-04T17:00:01Z"
                    messages:
                      - id: "01K5T0Q8C4000000000000M006"
                        level: info
                        title: Till released
                        text: Enter code SSP-TRF-8M3K-2QXD on the new PC within 7 days. Take a backup first.
                        showFromUtc: null
                        showUntilUtc: "2026-10-11T17:00:00Z"
                        dismissible: false
                        link: null
        "400":
          $ref: "#/components/responses/E400"
        "401":
          $ref: "#/components/responses/E401"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
...
  /api/v1/licence/validate:
    post:
      tags:
        - Licence
      operationId: validateLicence
      summary: Daily licence check, per till (sections 17.15.2, 17.5)
      description: "What the till sends (section 17.15.2, example perTill): EVERY till with a portal licence sends it for itself once onlineCheck.intervalHours have passed since its last successful check (a 15-minute timer), at start-up when due, and on Check again - identified by installId + licenceId + tokenSha256, no branch key required (sent when held). An unanswered check is repeated at the next tick with the same Idempotency-Key and the first attempt's body (tillClockUtc, clockWatermarkUtc included). The branch-model body (example leeds: the main till listing registers[], lastSyncAt, diagnostics) is not sent by this till version (section 17.5). Licence problems are 200 + status (active, expiring, expired, revoked, suspended, released; unknown = not active), never an error status. Only active and expiring let a per-till till trade: seatLimit (branch model, examples seatLimit and the registers[] seats) is not used for per-till licences and, if sent, locks it like any other status (section 17.9). The till acts on status, licenceToken, portalTimeUtc, messages and apiKey / hubUrl; licence, registers, minimumAppVersion and nextCheckAfterSeconds are not acted on yet."
      security:
        - {}
        - branchKey: []
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./licensing/schemas/validate-request.schema.json"
            examples:
              perTill:
                summary: "What the till sends: Till 1 checks in for itself"
                externalValue: "./licensing/samples/validate-request.per-till.json"
              leeds:
                summary: "Branch model (reference - not sent by this till version): Leeds, two tills"
                value:
                  licenceId: "01K5T0Q8C4000000000000Y001"
                  tokenSha256: "3848315811206cf590309eeefb2ce627a927f98167825e44049ea59910a2a3eb"
                  appVersion: "3.0.412"
                  approverKids:
                    - k76b44696
                  trustedKids:
                    - k13799fa4
                    - k76b44696
                  tillClockUtc: "2026-09-27T09:05:00Z"
                  clockWatermarkUtc: "2026-09-27T09:05:00Z"
                  lastValidatedAtUtc: "2026-09-26T09:00:05Z"
                  lastSyncAt: "2026-09-27T09:04:30Z"
                  lock:
                    locked: false
                    reason: null
                  registers:
                    - registerId: "01K5T0Q8C4000000000000R001"
                      code: "01"
                      name: Till 1
                      isMain: true
                      appVersion: "3.0.412"
                      installId: "01K5T0Q8C4000000000000H001"
                      deviceName: KIRKGATE-TILL1
                      lastSeenUtc: "2026-09-27T09:04:58Z"
                      isActive: true
                    - registerId: "01K5T0Q8C4000000000000R002"
                      code: "02"
                      name: Till 2
                      isMain: false
                      appVersion: "3.0.412"
                      installId: "01K5T0Q8C4000000000000H002"
                      deviceName: KIRKGATE-TILL2
                      lastSeenUtc: "2026-09-27T08:59:40Z"
                      isActive: true
                  diagnostics:
                    pendingSyncRows: 0
                    lastSyncError: null
                    databaseSizeMb: 84
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./licensing/schemas/validate-reply.schema.json"
              examples:
                perTill:
                  summary: "Per till: active, no new token"
                  externalValue: "./licensing/samples/validate-reply.per-till.json"
                released:
                  summary: "Per till: the key was released from this PC on the portal - the till locks after the sale"
                  externalValue: "./licensing/samples/validate-reply.released.json"
                trialExpiring:
                  summary: Trial, ends in 6 days
                  value:
                    status: expiring
                    licenceToken: null
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y001"
                      kind: trial
                      source: portal
                      expiresAt: "2026-10-03T09:00:00Z"
                      maxRegisters: 2
                      features: null
                      limits: null
                    registers:
                      - registerId: "01K5T0Q8C4000000000000R001"
                        seat: allowed
                      - registerId: "01K5T0Q8C4000000000000R002"
                        seat: allowed
                    apiKey: null
                    minimumAppVersion: null
                    portalTimeUtc: "2026-09-27T09:05:01Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M002"
                        level: warning
                        title: Trial ends soon
                        text: Your trial ends on 3 October 2026. Choose a plan to keep trading.
                        showFromUtc: null
                        showUntilUtc: "2026-10-03T09:00:00Z"
                        dismissible: true
                        link:
                          label: Choose a plan
                          url: "https://portal.example.co.uk/billing"
                renewed:
                  summary: "Paid: new token"
                  value:
                    status: active
                    licenceToken: SSPOS1.eyJ2IjoxLCJraWQiOiJrMTM3OTlmYTQiLCJsaWNlbmNlSWQiOiIwMUs1VDBROEM0MDAwMDAwMDAwMDAwWTAwMSIsImtpbmQiOiJmdWxsIiwic291cmNlIjoicG9ydGFsIiwiaXNzdWVyIjoiU1NQT1MgUG9ydGFsIiwiY29tcGFueUlkIjoiMDFLNVQwUThDNDAwMDAwMDAwMDAwMEMwMDEiLCJicmFuY2hJZCI6IjAxSzVUMFE4QzQwMDAwMDAwMDAwMDBCMDAxIiwiYnVzaW5lc3NOYW1lIjoiS2lya2dhdGUgU3RvcmVzIEx0ZCIsImJyYW5jaE5hbWUiOiJMZWVkcyBLaXJrZ2F0ZSIsIm1heFJlZ2lzdGVycyI6MywiaXNzdWVkQXQiOiIyMDI2LTEwLTAyVDA5OjA1OjAwWiIsInZhbGlkRnJvbSI6IjIwMjYtMTAtMDJUMDk6MDU6MDBaIiwiZXhwaXJlc0F0IjoiMjAyNy0xMC0wMlQyMzo1OTo1OVoiLCJvbmxpbmVDaGVjayI6eyJyZXF1aXJlZCI6dHJ1ZSwiaW50ZXJ2YWxIb3VycyI6MjQsImdyYWNlRGF5cyI6MTR9LCJub3RlcyI6IkFubnVhbCBwbGFuLCAzIHRpbGxzIiwiZmVhdHVyZXMiOlsid2hhdHNhcHAiLCJtdWx0aV9icmFuY2giXSwibGltaXRzIjp7InVzZXJzIjoxMCwiYnJhbmNoZXMiOjJ9fQ.NKDHN7Q69KhiEHCWFyDK-SKk8MUH_ESBp9yWyeryAgQpfByutZktbUAGR8ts2c5l34eCIIqnWscrKXN4ZeZFDA
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y001"
                      kind: full
                      source: portal
                      expiresAt: "2027-10-02T23:59:59Z"
                      maxRegisters: 3
                      features:
                        - whatsapp
                        - multi_branch
                      limits:
                        users: 10
                        branches: 2
                    registers:
                      - registerId: "01K5T0Q8C4000000000000R001"
                        seat: allowed
                      - registerId: "01K5T0Q8C4000000000000R002"
                        seat: allowed
                    apiKey: null
                    minimumAppVersion: null
                    portalTimeUtc: "2026-10-02T09:05:01Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M003"
                        level: info
                        title: Thank you
                        text: Your annual plan is active until 2 October 2027.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: true
                        link: null
                revoked:
                  summary: "Revoked: lock"
                  value:
                    status: revoked
                    licenceToken: null
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y001"
                      kind: trial
                      source: portal
                      expiresAt: "2026-10-03T09:00:00Z"
                      maxRegisters: 2
                      features: null
                      limits: null
                    registers:
                      - registerId: "01K5T0Q8C4000000000000R001"
                        seat: allowed
                      - registerId: "01K5T0Q8C4000000000000R002"
                        seat: allowed
                    apiKey: null
                    minimumAppVersion: null
                    portalTimeUtc: "2026-09-29T09:05:01Z"
                    nextCheckAfterSeconds: 3600
                    messages:
                      - id: "01K5T0Q8C4000000000000M004"
                        level: critical
                        title: Licence stopped
                        text: This licence has been stopped. Please call SSPOS support on 0113 496 0999.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: false
                        link: null
                seatLimit:
                  summary: "Too many tills: Till 3 refused"
                  value:
                    status: seatLimit
                    licenceToken: null
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y001"
                      kind: trial
                      source: portal
                      expiresAt: "2026-10-03T09:00:00Z"
                      maxRegisters: 2
                      features: null
                      limits: null
                    registers:
                      - registerId: "01K5T0Q8C4000000000000R001"
                        seat: allowed
                      - registerId: "01K5T0Q8C4000000000000R002"
                        seat: allowed
                      - registerId: "01K5T0Q8C4000000000000R004"
                        seat: refused
                    apiKey: null
                    minimumAppVersion: null
                    portalTimeUtc: "2026-09-28T09:05:01Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M005"
                        level: warning
                        title: Too many tills
                        text: Your plan allows 2 tills and 3 are in use. Till 3 is locked until you add a till to your plan or remove one.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: false
                        link:
                          label: Add a till
                          url: "https://portal.example.co.uk/billing/tills"
        "400":
          $ref: "#/components/responses/E400"
        "401":
          $ref: "#/components/responses/E401"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
  /api/v1/licence/redeem:
    post:
      tags:
        - Licence
      operationId: redeemLicence
      summary: "A key typed at a linked till, or a local key reported by a local till (sections 17.6, 17.16)"
      description: "(a) A key typed or pasted at a linked till (branch key). (b) Since 2026-09-28 a till holding a local (licence generator) key reports it once, when a licence server is set and reachable: keyType token, installCode, installId, no Authorization - the signed token is the credential. First sighting of a licenceId -> record licenceId -> installCode, 200 result recorded; the same licenceId from another installCode -> 409 key.used_on_another_install (the till locks). Renewals for that till are bound to its installCode. The body also carries deviceName, appVersion, os and tillClockUtc; a retry (the next daily check for a report) repeats the Idempotency-Key and the first body until you answer. Only 200 and 409 key.used_on_another_install change anything on the till."
      security:
        - branchKey: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Contract"
        - $ref: "#/components/parameters/AppVersion"
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/RegisterId"
        - $ref: "#/components/parameters/InstallId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              $ref: "./licensing/schemas/redeem-request.schema.json"
            examples:
              portalKey:
                summary: Key bought on the portal
                externalValue: "./licensing/samples/redeem-request.portal-key.json"
              localToken:
                summary: Token made by the licence generator
                externalValue: "./licensing/samples/redeem-request.local-token.json"
              localReport:
                summary: "Local till reports its open local key (no Authorization)"
                externalValue: "./licensing/samples/redeem-request.local-report.json"
      responses:
        "200":
          description: OK
          headers:
            X-SSPOS-Contract:
              $ref: "#/components/headers/Contract"
          content:
            "application/json":
              schema:
                $ref: "./licensing/schemas/redeem-reply.schema.json"
              examples:
                applied:
                  summary: New token to store
                  value:
                    result: applied
                    status: active
                    licenceToken: SSPOS1.eyJ2IjoxLCJraWQiOiJrMTM3OTlmYTQiLCJsaWNlbmNlSWQiOiIwMUs1VDBROEM0MDAwMDAwMDAwMDAwWTAwMSIsImtpbmQiOiJmdWxsIiwic291cmNlIjoicG9ydGFsIiwiaXNzdWVyIjoiU1NQT1MgUG9ydGFsIiwiY29tcGFueUlkIjoiMDFLNVQwUThDNDAwMDAwMDAwMDAwMEMwMDEiLCJicmFuY2hJZCI6IjAxSzVUMFE4QzQwMDAwMDAwMDAwMDBCMDAxIiwiYnVzaW5lc3NOYW1lIjoiS2lya2dhdGUgU3RvcmVzIEx0ZCIsImJyYW5jaE5hbWUiOiJMZWVkcyBLaXJrZ2F0ZSIsIm1heFJlZ2lzdGVycyI6MywiaXNzdWVkQXQiOiIyMDI2LTEwLTAyVDA5OjA1OjAwWiIsInZhbGlkRnJvbSI6IjIwMjYtMTAtMDJUMDk6MDU6MDBaIiwiZXhwaXJlc0F0IjoiMjAyNy0xMC0wMlQyMzo1OTo1OVoiLCJvbmxpbmVDaGVjayI6eyJyZXF1aXJlZCI6dHJ1ZSwiaW50ZXJ2YWxIb3VycyI6MjQsImdyYWNlRGF5cyI6MTR9LCJub3RlcyI6IkFubnVhbCBwbGFuLCAzIHRpbGxzIiwiZmVhdHVyZXMiOlsid2hhdHNhcHAiLCJtdWx0aV9icmFuY2giXSwibGltaXRzIjp7InVzZXJzIjoxMCwiYnJhbmNoZXMiOjJ9fQ.NKDHN7Q69KhiEHCWFyDK-SKk8MUH_ESBp9yWyeryAgQpfByutZktbUAGR8ts2c5l34eCIIqnWscrKXN4ZeZFDA
                    licence:
                      licenceId: "01K5T0Q8C4000000000000Y001"
                      kind: full
                      source: portal
                      expiresAt: "2027-10-02T23:59:59Z"
                      maxRegisters: 3
                      features:
                        - whatsapp
                        - multi_branch
                      limits:
                        users: 10
                        branches: 2
                    portalTimeUtc: "2026-10-02T09:05:01Z"
                    nextCheckAfterSeconds: 86400
                    messages:
                      - id: "01K5T0Q8C4000000000000M003"
                        level: info
                        title: Thank you
                        text: Your annual plan is active until 2 October 2027.
                        showFromUtc: null
                        showUntilUtc: null
                        dismissible: true
                        link: null
                localReport:
                  summary: "Local key report: first sighting recorded (licenceId -> installCode)"
                  externalValue: "./licensing/samples/redeem-reply.local-report.json"
                recorded:
                  summary: Local token recorded
                  value:
                    result: recorded
                    status: active
                    licenceToken: null
                    licence:
                      licenceId: "01K5M1GR8T000000000000Y002"
                      kind: full
                      source: local
                      expiresAt: "2027-03-31T23:59:59Z"
                      maxRegisters: 2
                      features:
                        - whatsapp
                      limits: null
                    portalTimeUtc: "2026-09-26T10:15:00Z"
                    nextCheckAfterSeconds: 86400
                    messages: []
        "400":
          $ref: "#/components/responses/E400"
        "401":
          $ref: "#/components/responses/E401"
        "403":
          $ref: "#/components/responses/E403"
        "404":
          $ref: "#/components/responses/E404"
        "409":
          $ref: "#/components/responses/E409"
        "410":
          $ref: "#/components/responses/E410"
        "422":
          $ref: "#/components/responses/E422"
        "426":
          $ref: "#/components/responses/E426"
        "429":
          $ref: "#/components/responses/E429"
        "500":
          $ref: "#/components/responses/E500"
        "503":
          $ref: "#/components/responses/E503"
...
components:
  securitySchemes:
    branchKey:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: "The branch's sync key (sync.hub_api_key): typed by the owner at Connect, or returned by the portal as apiKey in licence/activate, licence/validate or cloud/migrate. One key per branch. Required on sync/* and cloud/migrate/complete; sent on licence/* and devices/deactivate only when the till holds one."
  parameters:
    Contract:
      name: X-SSPOS-Contract
      in: header
      required: true
      description: Contract major version the till speaks. 1 today. See section 17.11.
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1"
    AppVersion:
      name: X-SSPOS-App-Version
      in: header
      required: true
      description: Till version.
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/appVersion"
      example: "3.0.412"
...
    CompanyId:
      name: X-SSPOS-Company-Id
      in: header
      required: true
      description: Company of the calling till.
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000C001"
    BranchId:
      name: X-SSPOS-Branch-Id
      in: header
      required: true
      description: "Branch of the calling till; must be the API key's branch."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000B001"
    RegisterId:
      name: X-SSPOS-Register-Id
      in: header
      required: true
      description: "The calling till: on sync and migration calls always the branch's main till (the sender); on licence and device calls the till licensing or releasing itself."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000R001"
    InstallId:
      name: X-SSPOS-Install-Id
      in: header
      required: true
      description: "The PC's install id (same as installId in the body). Sent on licence, device and migration calls; the sync calls (hello, push, pull) do not send it."
      schema:
        $ref: "./licensing/schemas/common.schema.json#/$defs/ulid"
      example: "01K5T0Q8C4000000000000H001"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: "One per logical request (the till sends a ULID). The till keeps the key from the first attempt until the portal answers and repeats it - with the first attempt's body byte for byte, tillClockUtc included - after no reply, 408, 409 request.in_progress, 429, 502, 503 or 504, across restarts, for up to 12 hours; any other reply ends the request. The portal stores (key, endpoint, caller, body hash) -> reply for at least 24 h and replays it for the same body; a different body -> 422 request.idempotency_mismatch; still running -> 409 request.in_progress + Retry-After. Never store a 408, 409 request.in_progress, 429, 502, 503 or 504 reply under the key (section 17.11 rule 5)."
      schema:
        type: string
        minLength: 8
        maxLength: 64
      example: "01K5T0Q8C4000000000000Z001"
...
  headers:
...
    Contract:
      description: Contract major version of this reply.
      schema:
        type: string
        pattern: "^[0-9]+$"
      example: "1"
  schemas:
    ErrorReply:
      $ref: "./licensing/schemas/error-reply.schema.json"
...
  responses:
    E400:
      description: "Bad request. Codes: `request.invalid` (*). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            request-invalid:
              summary: request.invalid
              value:
                code: request.invalid
                message: Body or query does not match the schema (missing field, bad ULID, bad date).
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  field: name of the first bad field
    E401:
      description: "Unauthorised. Codes: `auth.invalid_key` (sync/*, licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete); `auth.key_revoked` (sync/*, licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            sync-401:
              summary: samples/error-reply.401.json
              value:
                code: auth.invalid_key
                message: "The portal API key is not recognised. Check Settings > Hub sync > Portal API key."
                traceId: req-7c1f9a
                retryAfterSeconds: null
                rejectedKey: null
    E403:
      description: "Forbidden. Codes: `auth.wrong_branch` (sync/*, licence/validate, licence/redeem, devices/deactivate); `device.not_main_till` (licence/validate, licence/redeem, devices/deactivate, cloud/migrate/complete); `licence.seat_limit` (devices/activate, cloud/migrate, licence/activate); `licence.not_active` (devices/activate, cloud/migrate, licence/activate); `key.not_for_this_branch` (licence/redeem); `key.not_allowed` (licence/redeem). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            seat-limit:
              summary: error.seat-limit.403
              value:
                code: licence.seat_limit
                message: Your plan allows 2 tills and 2 are in use. Add a till to your plan on the portal.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details:
                  maxRegisters: 2
                  registersInUse: 2
    E404:
      description: "Not found. Codes: `activation.code_not_found` (devices/activate, cloud/migrate); `device.not_found` (devices/deactivate); `key.not_found` (licence/redeem, licence/activate, licence/validate); `migrate.upload_not_found` (sync/push (initial), cloud/migrate/complete); `sync_key.not_found` (cloud/migrate). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            activation-code-not-found:
              summary: error.activation-code-not-found.404
              value:
                code: activation.code_not_found
                message: That activation code is not recognised. Check the code in your welcome email.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details: null
...
    E422:
      description: "Unprocessable. Codes: `request.idempotency_mismatch` (*); `row.invalid` (sync/push); `licence.format` (licence/redeem, devices/activate, cloud/migrate); `licence.bad_signature` (licence/redeem, devices/activate, cloud/migrate); `licence.bad_signer_certificate` (licence/redeem, devices/activate, cloud/migrate); `licence.unknown_key` (licence/redeem, devices/activate, cloud/migrate); `licence.expired` (licence/redeem, cloud/migrate); `licence.wrong_shop` (licence/redeem, cloud/migrate); `licence.unsupported_version` (licence/redeem, devices/activate, cloud/migrate); `licence.wrong_branch` (licence/redeem). Full list: licensing/samples/error-codes.json."
      headers:
        X-SSPOS-Contract:
          $ref: "#/components/headers/Contract"
      content:
        "application/json":
          schema:
            $ref: "#/components/schemas/ErrorReply"
          examples:
            idempotency-mismatch:
              summary: "request.idempotency_mismatch - the same Idempotency-Key with a different body (the till never does this; it is a bug to report)"
              value:
                code: request.idempotency_mismatch
                message: This request key was already used with a different request.
                traceId: req-8a21f2
                retryAfterSeconds: null
                rejectedKey: null
                details: null
            licence-bad-signature:
              summary: error.licence-bad-signature.422
              value:
                code: licence.bad_signature
                message: This licence key is not genuine or has been changed. Paste it again or ask your dealer.
                traceId: req-8a21f0
                retryAfterSeconds: null
                rejectedKey: null
                details: null
            sync-422:
              summary: samples/error-reply.422.json
              value:
                code: row.invalid
                message: Sale 01K5VB000000000SR001000482 has a total with more than 2 decimal places.
                traceId: req-7c1fa0
                retryAfterSeconds: null
                rejectedKey: "Sale:01K5VB000000000SR001000482:1"
...